스트리밍 서비스 - 동영상, 광고 조회수, 동영상 중단

+++ src/main/java/com/sparta/binplay/dto/request/StreamRequestDto.java
@@ -6,12 +6,12 @@ import lombok.Getter;
 
 @Getter
 public class StreamRequestDto {
-    private Long viewingTime;
-    private Integer pausedAt;
+    private int viewingTime;
+    private int pausedAt;
     private Users user;
     private Videos video;
 
-    public StreamRequestDto(Long viewingTime, Integer pausedAt, Users user, Videos video) {
+    public StreamRequestDto(int viewingTime, int pausedAt, Users user, Videos video) {
         this.viewingTime = viewingTime;
         this.pausedAt = pausedAt;
         this.user = user;
